🚚 rename utils to kebab-case
