+++ packages/kitchn/src/components/Avatar/index.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 
 import { handlePixelValue, withDecorator } from "../../hoc/withDecorator";
 import { KitchnComponent } from "../../types";
-import { shortenName } from "../../utils/shortenName";
+import { shortenName } from "../../utils/shorten-name";
 import Text from "../Text";
 
 type Props = {
