+++ packages/kitchn/src/components/Button/index.tsx
@@ -4,8 +4,8 @@ import styled, { useTheme } from "styled-components";
 import { withDecorator } from "../../hoc/withDecorator";
 import { KitchnComponent, NormalSizes } from "../../types";
 import { AccentColors } from "../../types/theme";
-import { convertRGBToRGBA } from "../../utils/convertRGBToRGBA";
-import { isNumber } from "../../utils/isNumber";
+import { convertRGBToRGBA } from "../../utils/convert-rgb-to-rgba";
+import { isNumber } from "../../utils/is-number";
 import Spinner from "../Spinner";
 
 type Props = {
