+++ packages/kitchn/src/components/Checkbox/index.tsx
@@ -4,7 +4,7 @@ import styled from "styled-components";
 
 import { withDecorator } from "../../hoc/withDecorator";
 import { KitchnComponent } from "../../types";
-import { convertRGBToRGBA } from "../../utils/convertRGBToRGBA";
+import { convertRGBToRGBA } from "../../utils/convert-rgb-to-rgba";
 import Icon from "../Icon";
 
 type Props = {
