+++ packages/kitchn/src/components/Input/index.tsx
@@ -5,8 +5,8 @@ import styled from "styled-components";
 import { withDecorator } from "../../hoc/withDecorator";
 import { KitchnComponent, NormalSizes } from "../../types";
 import { AccentColors } from "../../types/theme";
-import { convertRGBToRGBA } from "../../utils/convertRGBToRGBA";
-import { isNumber } from "../../utils/isNumber";
+import { convertRGBToRGBA } from "../../utils/convert-rgb-to-rgba";
+import { isNumber } from "../../utils/is-number";
 import Container, { ContainerProps } from "../Container";
 import Error from "../Error";
 import Icon, { IconProps } from "../Icon";
