+++ packages/kitchn/src/native/components/Input/index.tsx
@@ -10,8 +10,8 @@ import {
 import styled, { useTheme } from "styled-components/native";
 
 import { capitalize } from "../../../utils/capitalize";
-import { convertRGBToRGBA } from "../../../utils/convertRGBToRGBA";
-import { isNumber } from "../../../utils/isNumber";
+import { convertRGBToRGBA } from "../../../utils/convert-rgb-to-rgba";
+import { isNumber } from "../../../utils/is-number";
 import { withScale } from "../../hoc/withScale";
 import { AccentColors, KitchnComponent, NormalSizes } from "../../types";
 import Icon, { IconProps } from "../Icon";
