+++ packages/kitchn/src/utils/_collections.ts
@@ -1,8 +1,8 @@
 export * from "./capitalize";
-export * from "./convertRGBToRGBA";
-export * from "./getId";
-export * from "./isNumber";
-export * from "./isString";
-export * from "./isDevelopment";
-export * from "./pickChild";
-export * from "./shortenName";
+export * from "./convert-rgb-to-rgba";
+export * from "./get-id";
+export * from "./is-number";
+export * from "./is-string";
+export * from "./is-development";
+export * from "./pick-child";
+export * from "./shorten-name";
