+++ packages/kitchn/src/utils/_layout.ts
@@ -1,10 +1,10 @@
 export * from "./codes";
-export * from "./getActiveModMap";
-export * from "./getCtrlKeysByPlatform";
-export * from "./getElementOffset";
-export * from "./getTranslateByPlacement";
-export * from "./isBrowser";
-export * from "./isMac";
-export * from "./isRefTarget";
+export * from "./get-active-mod-map";
+export * from "./get-ctrl-keys-by-platform";
+export * from "./get-element-offset";
+export * from "./get-translate-by-placement";
+export * from "./is-browser";
+export * from "./is-mac";
+export * from "./is-ref-target";
 export * from "./tooltip";
 export * from "./animate";
