+++ packages/kitchn/src/utils/get-ctrl-keys-by-platform.ts
@@ -1,4 +1,4 @@
-import { isMac } from "./isMac";
+import { isMac } from "./is-mac";
 
 export const getCtrlKeysByPlatform = (): Record<
   string,
