+++ packages/kitchn/src/utils/is-browser.ts
