+++ packages/kitchn/src/utils/is-mac.ts
@@ -1,4 +1,4 @@
-import { isBrowser } from "./isBrowser";
+import { isBrowser } from "./is-browser";
 
 export const isMac = (): boolean => {
   if (!isBrowser()) return false;
