+++ packages/kitchn/src/utils/is-number.ts
