+++ packages/kitchn/src/utils/is-string.ts
