+++ packages/kitchn/src/utils/shorten-name.ts
